% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcPIPs.R
\name{CalcGroupPIPs}
\alias{CalcGroupPIPs}
\title{Calculate group-specific posterior inclusion probabilities}
\usage{
CalcGroupPIPs(fit, sel = NULL)
}
\arguments{
\item{fit}{}

\item{sel}{}
}
\description{
Calculate posterior inclusion probabilities for each group of pollutants
}

