% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalcPIPs.R
\name{CalcPIPs}
\alias{CalcPIPs}
\title{Calculate pollutant-specific posterior inclusion probabilities}
\usage{
CalcPIPs(fit, sel = NULL)
}
\arguments{
\item{fit}{}

\item{sel}{}
}
\description{
Calculate pollutant-specific posterior inclusion probabilities
}

