% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPIPs.R
\name{CalcWithinGroupPIPs}
\alias{CalcWithinGroupPIPs}
\title{Calculate conditional predictor specific posterior inclusion probabilities}
\usage{
CalcWithinGroupPIPs(fit, sel = NULL)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}
}
\description{
For those predictors within a multi-preditor group, as defined using the \code{groups} argument, the posterior inclusion probabilities for the predictor conditional on the group being selected into the model.
}
