% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComputePostmeanHnew.R
\name{ComputePostmeanHnew}
\alias{ComputePostmeanHnew}
\title{Compute the posterior mean and variance of \code{h} at a new predictor values}
\usage{
ComputePostmeanHnew(fit, y, Z, X, Znew, sel = NULL)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{Znew}{}

\item{sel}{A vector selecting which iterations of the BKMR fit should be retained for inference. Currently only implemented for \code{method == "r"}.}
}
\description{
Function to approximate the posterior mean and variance as a function of the estimated tau, lambda, beta, and sigsq.eps
}

