% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExposureResponseFunctions.R
\name{ExposureResponseBivarLevels}
\alias{ExposureResponseBivarLevels}
\title{Function to plot the exposure response function of a particular pollutant at different levels (quantiles) of a second pollutant}
\usage{
ExposureResponseBivarLevels(expos.resp.df, expos, qs = c(0.25, 0.5, 0.75))
}
\arguments{
\item{expos.resp.df}{object obtained from running the function \code{ExposureResponseBivar()}}

\item{qs}{vector of quantiles of the second pollutant}
}
\description{
Function to plot the exposure response function of a particular pollutant at different levels (quantiles) of a second pollutant
}

