% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExtractEsts.R
\name{ExtractEsts}
\alias{ExtractEsts}
\title{Extract summary statistics}
\usage{
ExtractEsts(fit, q = c(0.025, 0.25, 0.5, 0.75, 0.975), sel = NULL)
}
\description{
Obtain summary statistics of each parameter from the BKMR fit
}

