% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExtractEsts.R
\name{ExtractSamps}
\alias{ExtractSamps}
\title{Extract samples of each parameter from the BKMR fit}
\usage{
ExtractSamps(fit, sel = NULL)
}
\description{
Extract samples of each parameter from the BKMR fit
}

