% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KMR.R
\name{FitKMR}
\alias{FitKMR}
\title{Kernel machine regression (KMR) with Gaussian kernel}
\usage{
FitKMR(y, Z, X, max.iter = 500, tol = 1e-05)
}
\arguments{
\item{y}{}

\item{Z}{}

\item{X}{}

\item{max.iter}{}
}
\description{
The rho parameter is also estimated (rather than fixed)
}

