% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InvestigatePrior_rm.R
\name{InvestigatePrior}
\alias{InvestigatePrior}
\title{Investigate prior}
\usage{
InvestigatePrior(y, Z, X, ngrid = 50, q.seq = c(2, 1, 1/2, 1/4, 1/8, 1/16),
  verbose = FALSE)
}
\arguments{
\item{ngrid}{Number of grid points over which to plot the exposure-response function}

\item{q.seq}{Sequence of values corresponding to different degrees of smoothness in the estimated exposure-response function. A value of q corresponds to fractions of the range of the data over which there is a decay in the correlation \code{cor(h[i],h[j])} between two subjects by 50\code{\%}.}

\item{verbose}{TRUE or FALSE: flag indicating whether to print to the screen which pollutant and q value has been completed}
}
\description{
Investigate impact of prior distribution for \code{r[m]} parameters on smoothness of exposure-response function \code{h()}.
}

