% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{OverallRiskSummaries}
\alias{OverallRiskSummaries}
\title{Calculate overall risk summaries}
\usage{
OverallRiskSummaries(fit, y = NULL, Z = NULL, X = NULL, qs = seq(0.25,
  0.75, by = 0.05), q.fixed = 0.5, method = "approx", sel = NULL)
}
\arguments{
\item{y}{a vector of outcome data of length \code{n}.}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{qs}{vector of quantiles at which to calculate the overall risk summary}

\item{q.fixed}{a second quantile at which to compare the estimated \code{h} function}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets; see details}

\item{sel}{selects which iterations of the MCMC sampler to use for inference; see details}
}
\description{
Compare estimated \code{h} function when all predictors are at a particular quantile to when all are at a second fixed quantile
}
\details{
see \code{\link{ComputePostmeanHnew}} for details.
}

