% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OverallTest.R
\name{OverallTest}
\alias{OverallTest}
\title{Overall kernel test}
\usage{
OverallTest(y, Z, X = NULL, kernel = "gaussian", rho = 1, d = 2)
}
\arguments{
\item{y}{}

\item{Z}{}

\item{X}{}

\item{kernel}{}

\item{rho}{tuning parameter for various kernel functions}

\item{d}{tuning parameter for dth polynomial kernel}
}
\description{
Conduct an overall test of the mixture using the approach of Liu, Lin, and Ghosh (2007) Biometrics.
}

