% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCorr.R
\name{PlotCorr}
\alias{PlotCorr}
\title{Plot of correlation matrix}
\usage{
PlotCorr(Z, reorder = TRUE, print.vals = FALSE, plot = TRUE, digits = 2)
}
\arguments{
\item{Z}{}

\item{reorder}{order correlation matrix based on distance}

\item{print.vals}{show the correlation values on the plto}

\item{plot}{make the plot}

\item{digits}{If the correlation values are to be printed in the plot (\code{vals == TRUE}), how many digits should be displayed}
}
\description{
Plot of correlation matrix
}

