% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvestigatePrior_rm.R
\name{PlotPriorFits}
\alias{PlotPriorFits}
\title{PlotPriorFits}
\usage{
PlotPriorFits(y, X, Z, fits, which.z = NULL, which.q = NULL,
  plot.resid = TRUE, ylim = NULL, ...)
}
\arguments{
\item{y}{a vector of outcome data of length \code{n}.}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{...}{}
}
\value{
fitted objects
}
\description{
PlotPriorFits
}

