% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{PolRiskSummary}
\alias{PolRiskSummary}
\title{Compare estimated exposure-response function when a single pollutant (or a set of pollutants) is at the 75th versus 25th percentile, when all of the other exposures are fixed at a particular percentile}
\usage{
PolRiskSummary(whichpol = 1, fit, y, expos, covar, qs.diff = c(0.25, 0.75),
  q.fixed = 0.5, preds.method = "approx", ...)
}
\arguments{
\item{whichpol}{a scalar or vector selecting which pollutants to compute the summary for (the other pollutants will be fixed at the value \code{q.fixed})}
}
\description{
Compare estimated exposure-response function when a single pollutant (or a set of pollutants) is at the 75th versus 25th percentile, when all of the other exposures are fixed at a particular percentile
}

