% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PredictorResponseFunctions.R
\name{PredictorResponseBivar}
\alias{PredictorResponseBivar}
\title{Predict the exposure-response function at a new grid of points}
\usage{
PredictorResponseBivar(fit, y = NULL, Z = NULL, X = NULL,
  z.pairs = subset(expand.grid(z1 = 1:ncol(Z), z2 = 1:ncol(Z)), z1 < z2),
  preds.method = "approx", ngrid = 50, q.fixed = 0.5, sel = NULL,
  min.plot.dist = 0.5, center = TRUE, z.names = colnames(Z),
  verbose = TRUE, ...)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{z.pairs}{Data frame showing which pairs of pollutants to plot}

\item{preds.method}{}

\item{ngrid}{}

\item{q.fixed}{}

\item{sel}{}

\item{min.plot.dist}{}

\item{center}{}

\item{z.names}{}

\item{verbose}{TRUE or FALSE: flag of whether to print intermediate output to the screen}

\item{...}{}
}
\description{
Predict the exposure-response function at a new grid of points
}

