% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PredictorResponseFunctions.R
\name{PredictorResponseBivarLevels}
\alias{PredictorResponseBivarLevels}
\title{Function to plot the \code{h} function of a particular variable at different levels (quantiles) of a second variable}
\usage{
PredictorResponseBivarLevels(pred.resp.df, Z = NULL, qs = c(0.25, 0.5,
  0.75))
}
\arguments{
\item{pred.resp.df}{object obtained from running the function \code{PredictorResponseBivar()}}

\item{qs}{vector of quantiles of the second variable}
}
\description{
Function to plot the \code{h} function of a particular variable at different levels (quantiles) of a second variable
}

