% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictorResponseFunctions.R
\name{PredictorResponseBivarPair}
\alias{PredictorResponseBivarPair}
\title{PredictorResponseBivarPair}
\usage{
PredictorResponseBivarPair(fit, y, Z, X, whichz1 = 1, whichz2 = 2, whichz3,
  preds.method = "approx", prob = 0.5, q.fixed = 0.5, sel = NULL,
  ngrid = 50, min.plot.dist = 0.5, center = TRUE, ...)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{whichz1}{}

\item{whichz2}{}

\item{whichz3}{}

\item{preds.method}{}

\item{prob}{}

\item{q.fixed}{}

\item{sel}{}

\item{ngrid}{}

\item{min.plot.dist}{}

\item{center}{}

\item{...}{}
}
\description{
PredictorResponseBivarPair
}

