% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictorResponseFunctions.R
\name{PredictorResponseUnivar}
\alias{PredictorResponseUnivar}
\title{PredictorResponseUnivar}
\usage{
PredictorResponseUnivar(fit, y = NULL, Z = NULL, X = NULL,
  which.z = 1:ncol(Z), preds.method = "approx", ngrid = 50,
  q.fixed = 0.5, sel = NULL, min.plot.dist = Inf, center = TRUE,
  z.names = colnames(Z), ...)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{which.z}{}

\item{preds.method}{}

\item{ngrid}{}

\item{q.fixed}{}

\item{sel}{}

\item{min.plot.dist}{}

\item{center}{}

\item{z.names}{}
}
\description{
PredictorResponseUnivar
}

