% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PredictorResponseFunctions.R
\name{PredictorResponseUnivarVar}
\alias{PredictorResponseUnivarVar}
\title{PredictorResponseUnivarVar}
\usage{
PredictorResponseUnivarVar(whichz = 1, fit, y, Z, X,
  preds.method = "approx", ngrid = 50, q.fixed = 0.5, sel = NULL,
  min.plot.dist = Inf, center = TRUE, z.names = colnames(Z), ...)
}
\arguments{
\item{whichz}{}

\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{preds.method}{}

\item{ngrid}{}

\item{q.fixed}{}

\item{sel}{}

\item{min.plot.dist}{}

\item{center}{}

\item{z.names}{}

\item{...}{}
}
\description{
PredictorResponseUnivarVar
}

