% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr_main_functions.R
\name{print.bkmrfit}
\alias{print.bkmrfit}
\title{Print basic summary of BKMR model fit}
\usage{
\method{print}{bkmrfit}(x, digits = 5, ...)
}
\arguments{
\item{x}{an object of class "bkmrfit"}

\item{digits}{the number of digits to show when printing}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} method for class "bkmrfit"
}
