% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplePred.R
\name{SamplePred}
\alias{SamplePred}
\title{Obtain posterior samples of predictions at new points}
\usage{
SamplePred(fit, Znew, Xnew = NULL, Z = NULL, X = NULL, y = NULL,
  sel = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{Znew}{optional matrix of new predictor values at which to predict new \code{h}, where each row represents a new observation. This will slow down the model fitting.}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{y}{a vector of outcome data of length \code{n}.}

\item{sel}{A vector selecting which iterations of the BKMR fit should be retained for inference. If not specified, will default to keeping every 10 iterations after dropping the first 50% of samples, or if this results in fewer than 100 iterations, than 100 iterations are kept}
}
\description{
Obtains posterior samples of \code{E(Y) = h(Znew) + beta*Xnew} or of \code{g^{-1}[E(y)]}
}

