% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{SeqVarRiskSummaries}
\alias{SeqVarRiskSummaries}
\title{Sequentially add predictors to compute the overall risk, when the set of predictors in $Z$ are all at their 75th vs. 25th percentile, for all of the other predictors fixed at a particular percentile}
\usage{
SeqVarRiskSummaries(fit, y, Z, X, which.z = 1:ncol(Z), qs.diff = c(0.25,
  0.75), q.fixed = 0.5, preds.method = "approx", sel = NULL,
  z.names = colnames(Z), ...)
}
\description{
Sequentially add predictors to compute the overall risk, when the set of predictors in $Z$ are all at their 75th vs. 25th percentile, for all of the other predictors fixed at a particular percentile
}

