% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{Simulate dataset}
\usage{
SimData(n = 100, M = 5, sigsq.true = 0.5, beta.true = 2,
  GenCovar = NULL, GenExpos = NULL, HFun = NULL)
}
\arguments{
\item{n}{Number of observations}

\item{M}{Number of exposure variables to generate}

\item{sigsq.true}{Variance of normally distributed residual error}

\item{beta.true}{Coefficient on the covariate}

\item{GenCovar}{A function that takes as input a vector of length \code{M}. Defaults to \code{\link{GenCovarDefault}}.}

\item{GenExpos}{A function that takes as input a vector of length \code{M} Defaults to \code{\link{GenExposDefault}}.}

\item{HFun}{A function that takes as input a vector of length \code{M} and outputs a scalar}
}
\description{
Simulate dataset
}
\examples{
set.seed(5)
dat <- SimData()
}

