% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData1}
\alias{SimData1}
\title{Simulate dataset}
\usage{
SimData1(n = 100, M = 5, sigsq.true = 0.5, beta.true = 2,
  GenCovar = NULL, GenExpos = NULL, HFun = NULL)
}
\arguments{
\item{n}{}

\item{M}{Number of exposure variables to generate}

\item{sigsq.true}{Variance of normally distributed residual error}

\item{beta.true}{Coefficient on the covariate}

\item{GenCovar}{A function that takes as input a vector of length \code{M}. Defaults to \code\link{GenCovarDefault}.}

\item{GenExpos}{A function that takes as input a vector of length \code{M}}

\item{HFun}{A function that takes as input a vector of length \code{M} and outputs a scalar}
}
\description{
Simulate dataset
}
\examples{
set.seed(5)
dat <- SimData1()
}

