% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{SingPolRiskSummary}
\alias{SingPolRiskSummary}
\title{Compare estimated exposure-response function when a single pollutant is at the 75th versus 25th percentile, when all of the other exposures are fixed at a particular percentile}
\usage{
SingPolRiskSummary(whichpol = 1, fit, y, expos, covar, qs.diff = c(0.25,
  0.75), q.fixed = 0.5, preds.method = "approx", ...)
}
\description{
Compare estimated exposure-response function when a single pollutant is at the 75th versus 25th percentile, when all of the other exposures are fixed at a particular percentile
}

