% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{SingVarIntSummaries}
\alias{SingVarIntSummaries}
\title{SingVarIntSummaries}
\usage{
SingVarIntSummaries(fit, y, Z, X, which.z = 1:ncol(Z), qs.diff = c(0.25,
  0.75), qs.fixed = c(0.25, 0.75), preds.method = "approx", sel = NULL,
  z.names = colnames(Z), ...)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{which.z}{}

\item{qs.diff}{}

\item{qs.fixed}{}

\item{preds.method}{}

\item{sel}{}

\item{z.names}{}

\item{...}{}
}
\description{
SingVarIntSummaries
}

