% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{SingVarIntSummary}
\alias{SingVarIntSummary}
\title{Compare the single-predictor health risks when all of the other predictors in Z are fixed to their 75th percentile to when all of the other predictors in Z are fixed to their 25th percentile.}
\usage{
SingVarIntSummary(whichz = 1, fit, y, Z, X, qs.diff = c(0.25, 0.75),
  qs.fixed = c(0.25, 0.75), preds.method = "approx", sel = NULL, ...)
}
\description{
Compare the single-predictor health risks when all of the other predictors in Z are fixed to their 75th percentile to when all of the other predictors in Z are fixed to their 25th percentile.
}

