% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{SingVarRiskSummaries}
\alias{SingVarRiskSummaries}
\title{SingVarRiskSummaries}
\usage{
SingVarRiskSummaries(fit, y, Z, X, which.z = 1:ncol(Z), qs.diff = c(0.25,
  0.75), q.fixed = c(0.25, 0.5, 0.75), preds.method = "approx",
  sel = NULL, z.names = colnames(Z), ...)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{which.z}{}

\item{qs.diff}{}

\item{q.fixed}{}

\item{preds.method}{}

\item{sel}{}

\item{z.names}{}

\item{...}{}
}
\description{
SingVarRiskSummaries
}

