% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StanBKMR.R
\name{StanBKMR}
\alias{StanBKMR}
\title{Wrapper function to fit BKMR using Stan}
\usage{
StanBKMR(y, Z, X, file = NULL, ...)
}
\arguments{
\item{y}{a vector of outcome data of length \code{n}.}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{...}{}

\item{Optional}{path to the file containing Stan model code.}
}
\value{
Object of class \code{"\link[=stanfit-class]{stanfit}"}
}
\description{
Fits Bayesian kernel machine regression (BKMR) for cross-sectional data using \code{\link[rstan]{stan}}.
}
\examples{
\dontrun{
library(rstan)
set.seed(1)
dat <- SimData(n = 50, M = 4, sigsq.true = 0.5, beta.true = 2)
y <- dat$y
Z <- dat$Z
X <- dat$X
set.seed(111)
runtime <- system.time({
    fitstanex1 <- StanBKMR(y = y, Z = Z, X = X, iter = 400, chains = 3)
})
print(fitstanex1, par = c("beta", "sigma_sq", "h[1]", "h[100]", "lp__"))
}
}

