% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TracePlot.R
\name{TracePlot}
\alias{TracePlot}
\title{Trace plot}
\usage{
TracePlot(fit, par, comp = 1, sel = NULL, main = "", xlab = "iteration",
  ylab = "parameter value", ...)
}
\arguments{
\item{fit}{}

\item{par}{}

\item{comp}{}

\item{sel}{}

\item{main}{}

\item{xlab}{}

\item{ylab}{}

\item{...}{}
}
\description{
Trace plot
}

