% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{VarRiskSummary}
\alias{VarRiskSummary}
\title{Compare estimated \code{h} function when a single variable (or a set of variables) is at the 75th versus 25th percentile, when all of the other variables are fixed at a particular percentile}
\usage{
VarRiskSummary(whichz = 1, fit, y, Z, X, qs.diff = c(0.25, 0.75),
  q.fixed = 0.5, preds.method = "approx", sel = NULL, ...)
}
\arguments{
\item{whichz}{a scalar or vector selecting which Z variables to compute the summary for (the other variables in Z will be fixed at the value \code{q.fixed})}
}
\description{
Compare estimated \code{h} function when a single variable (or a set of variables) is at the 75th versus 25th percentile, when all of the other variables are fixed at a particular percentile
}

