% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPIPs.R
\name{CalcGroupPIPs}
\alias{CalcGroupPIPs}
\title{Calculate group-specific posterior inclusion probabilities}
\usage{
CalcGroupPIPs(fit, sel = NULL)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}
}
\description{
Calculate posterior inclusion probabilities for each group of variables
}
