% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPIPs.R
\name{CalcPIPs}
\alias{CalcPIPs}
\title{Calculate variable-specific posterior inclusion probabilities}
\usage{
CalcPIPs(fit, sel = NULL)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}
}
\description{
Calculate variable-specific posterior inclusion probabilities from BKMR model fit
}
