% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPIPs.R
\name{ExtractPIPs}
\alias{ExtractPIPs}
\title{Extract posterior inclusion probabilities (PIPs) from BKMR model fit}
\usage{
ExtractPIPs(fit, sel = NULL, z.names = NULL)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}

\item{z.names}{optional argument providing the names of the variables included in the \code{h} function.}
}
\value{
a data frame with the variable-specific PIPs for BKMR fit with component-wise variable selection, and with the group-specific and conditional (within-group) PIPs for BKMR fit with hierarchical variable selection.
}
\description{
Extract posterior inclusion probabilities (PIPs) from Bayesian Kernel Machine Regression (BKMR) model fit
}

