% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractEsts.R
\name{ExtractSamps}
\alias{ExtractSamps}
\title{Extract samples}
\usage{
ExtractSamps(fit, sel = NULL)
}
\arguments{
\item{fit}{An object containing the results returned by a the \code{kmbayes} function}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}
}
\description{
Extract samples of each parameter from the BKMR fit
}
