% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ComputeMixtureSummaries.R
\name{OverallRiskSummaries}
\alias{OverallRiskSummaries}
\title{OverallRiskSummaries}
\usage{
OverallRiskSummaries(fit, y, Z, X, qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5, preds.method = "approx", sel = NULL)
}
\arguments{
\item{fit}{}

\item{y}{}

\item{Z}{}

\item{X}{}

\item{qs}{}

\item{q.fixed}{}

\item{preds.method}{}

\item{sel}{}
}
\description{
Compare estimated \code{h} function when all predictors are at a particular percentile to when all are at the 50th percentile
}

