% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OverallTest.R
\name{OverallTest}
\alias{OverallTest}
\title{Conduct an overall test of the mixture using the approach of Liu, Lin, and Ghosh (2007) Biometrics.}
\usage{
OverallTest(y, expos, covar = NULL, kernel = "gaussian", rho = 1, d = 2)
}
\arguments{
\item{rho}{tuning parameter for various kernel functions}

\item{d}{tuning parameter for dth polynomial kernel}
}
\description{
Conduct an overall test of the mixture using the approach of Liu, Lin, and Ghosh (2007) Biometrics.
}

