% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotCorr.R
\name{PlotCorr}
\alias{PlotCorr}
\title{Plot of correlation matrix}
\usage{
PlotCorr(Z, reorder = TRUE, print.vals = FALSE, plot = TRUE, digits = 2)
}
\arguments{
\item{digits}{If the correlation values are to be printed in the plot (\code{vals == TRUE}), how many digits should be displayed}

\item{vals}{}
}
\description{
Plot of correlation matrix
}

