% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InvestigatePrior_rm.R
\name{PlotPriorFits}
\alias{PlotPriorFits}
\title{PlotPriorFits}
\usage{
PlotPriorFits(y, X, Z, fits, which.z = NULL, which.q = NULL,
  plot.resid = TRUE, ylim = NULL, ...)
}
\arguments{
\item{y}{}

\item{X}{}

\item{Z}{}

\item{fits}{}

\item{which.z}{}

\item{which.q}{}

\item{plot.resid}{}

\item{ylim}{}

\item{...}{}
}
\value{
fitted objects
}
\description{
PlotPriorFits
}

