% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/InvestigatePrior_rm.R
\name{PlotPriorFits}
\alias{PlotPriorFits}
\title{PlotPriorFits}
\usage{
PlotPriorFits(y, covar, expos, fits, which.expos = NULL, which.q = NULL,
  plot.resid = TRUE, ylim = NULL, ...)
}
\arguments{
\item{y}{}

\item{covar}{}

\item{expos}{}

\item{fits}{}

\item{which.expos}{}

\item{which.q}{}

\item{plot.resid}{}

\item{ylim}{}

\item{...}{}
}
\value{
fitted objects
}
\description{
PlotPriorFits
}

