% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictorResponseFunctions.R
\name{PredictorResponseBivarLevels}
\alias{PredictorResponseBivarLevels}
\title{Plot cross-sections of the bivariate predictor-response function}
\usage{
PredictorResponseBivarLevels(pred.resp.df, Z = NULL, qs = c(0.25, 0.5,
  0.75), both_pairs = TRUE)
}
\arguments{
\item{pred.resp.df}{object obtained from running the function \code{\link{PredictorResponseBivar}}}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{qs}{vector of quantiles at which to fix the second variable}

\item{both_pairs}{flag indicating whether, if \code{h(z1)} is being plotted for z2 fixed at different levels, that they should be plotted in the reverse order as well (for \code{h(z2)} at different levels of z1)}
}
\description{
Function to plot the \code{h} function of a particular variable at different levels (quantiles) of a second variable
}

