% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr_main_functions.R
\name{summary.bkmrfit}
\alias{summary.bkmrfit}
\title{Summarizing BKMR model fits}
\usage{
\method{summary}{bkmrfit}(object, q = c(0.025, 0.975), digits = 5,
  show_ests = TRUE, show_MH = TRUE, ...)
}
\arguments{
\item{object}{an object of class "bkmrfit"}

\item{q}{quantiles of posterior distribution to show}

\item{digits}{the number of digits to show when printing}

\item{show_ests}{logical; if \code{TRUE}, prints summary statistics of posterior distribution}

\item{show_MH}{logical; if \code{TRUE}, prints acceptance rates from the Metropolis-Hastings algorithm}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{summary} method for class "bkmrfit"
}
